requirements.txt
Pillow 
pandas
numpy 
requests
streamlit

